販売手数料・利益jsファイルを別ファイルへ移動
